refactor and create separate trait file for fetch methods
