+++ app/Http/Controllers/Admin/Traits/FetchAppraisalTypeTrait.php
@@ -1,0 +1,11 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchAppraisalTypeTrait
+{
+    public function fetchAppraisalType()
+    {
+        return $this->fetch(\App\Models\AppraisalType::class);
+    }
+}
