+++ app/Http/Controllers/Admin/Traits/FetchBloodTypeTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchBloodTypeTrait
+{
+    public function fetchBloodType()
+    {
+        return $this->fetch(\App\Models\BloodType::class);
+    }
+    
+}
