+++ app/Http/Controllers/Admin/Traits/FetchCitizenshipTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchCitizenshipTrait
+{
+    public function fetchCitizenship()
+    {
+        return $this->fetch(\App\Models\Citizenship::class);
+    }
+    
+}
