+++ app/Http/Controllers/Admin/Traits/FetchCivilStatusTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchCivilStatusTrait
+{
+    public function fetchCivilStatus()
+    {
+        return $this->fetch(\App\Models\CivilStatus::class);
+    }
+    
+}
