+++ app/Http/Controllers/Admin/Traits/FetchCompanyTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchCompanyTrait
+{
+    public function fetchCompany()
+    {
+        return $this->fetch(\App\Models\Company::class);
+    }
+    
+}
