+++ app/Http/Controllers/Admin/Traits/FetchDepartmentTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchDepartmentTrait
+{
+    public function fetchDepartment()
+    {
+        return $this->fetch(\App\Models\Department::class);
+    }
+    
+}
