+++ app/Http/Controllers/Admin/Traits/FetchDivisionTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchDivisionTrait
+{
+    public function fetchDivision()
+    {
+        return $this->fetch(\App\Models\Division::class);
+    }
+    
+}
