+++ app/Http/Controllers/Admin/Traits/FetchEducationalLevelTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchEducationalLevelTrait
+{
+    public function fetchEducationalLevel()
+    {
+        return $this->fetch(\App\Models\EducationalLevel::class);
+    }
+    
+}
