+++ app/Http/Controllers/Admin/Traits/FetchEmploymentstatusTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchEmploymentstatusTrait
+{
+    public function fetchEmploymentstatus()
+    {
+        return $this->fetch(\App\Models\EmploymentStatus::class);
+    }
+    
+}
