+++ app/Http/Controllers/Admin/Traits/FetchGenderTrait.php
@@ -1,0 +1,11 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchGenderTrait
+{
+    public function fetchGender()
+    {
+        return $this->fetch(\App\Models\Gender::class);
+    }
+}
