+++ app/Http/Controllers/Admin/Traits/FetchGravityOfSanctionTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchGravityOfSanctionTrait
+{
+    public function fetchGravityOfSanction()
+    {
+        return $this->fetch(\App\Models\GravityOfSanction::class);
+    }
+    
+}
