+++ app/Http/Controllers/Admin/Traits/FetchJobstatusTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchJobstatusTrait
+{
+    public function fetchJobstatus()
+    {
+        return $this->fetch(\App\Models\JobStatus::class);
+    }
+    
+}
