+++ app/Http/Controllers/Admin/Traits/FetchLevelTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchLevelTrait
+{
+    public function fetchLevel()
+    {
+        return $this->fetch(\App\Models\Level::class);
+    }
+    
+}
