+++ app/Http/Controllers/Admin/Traits/FetchLocationTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchLocationTrait
+{
+    public function fetchLocation()
+    {
+        return $this->fetch(\App\Models\Location::class);
+    }
+    
+}
