+++ app/Http/Controllers/Admin/Traits/FetchOffenceClassificationTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchOffenceClassificationTrait
+{
+    public function fetchOffenceClassification()
+    {
+        return $this->fetch(\App\Models\OffenceClassification::class);
+    }
+    
+}
