+++ app/Http/Controllers/Admin/Traits/FetchPaymentmethodTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchPaymentmethodTrait
+{
+    public function fetchPaymentmethod()
+    {
+        return $this->fetch(\App\Models\PaymentMethod::class);
+    }
+    
+}
