+++ app/Http/Controllers/Admin/Traits/FetchPositionTrait.php
@@ -1,0 +1,11 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchPositionTrait
+{
+    public function fetchPosition()
+    {
+        return $this->fetch(\App\Models\Position::class);
+    }
+}
