+++ app/Http/Controllers/Admin/Traits/FetchRankTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchRankTrait
+{
+    public function fetchRank()
+    {
+        return $this->fetch(\App\Models\Rank::class);
+    }
+    
+}
