+++ app/Http/Controllers/Admin/Traits/FetchReligionTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchReligionTrait
+{
+    public function fetchReligion()
+    {
+        return $this->fetch(\App\Models\Religion::class);
+    }
+    
+}
