+++ app/Http/Controllers/Admin/Traits/FetchSectionTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchSectionTrait
+{
+    public function fetchSection()
+    {
+        return $this->fetch(\App\Models\Section::class);
+    }
+    
+}
