+++ app/Http/Controllers/Admin/Traits/FetchShiftScheduleTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchShiftScheduleTrait
+{
+    public function fetchShiftSchedule()
+    {
+        return $this->fetch(\App\Models\ShiftSchedule::class);
+    }
+    
+}
