+++ app/Http/Controllers/Admin/Traits/FetchTeamTrait.php
@@ -1,0 +1,12 @@
+<?php 
+
+namespace App\Http\Controllers\Admin\Traits;
+
+trait FetchTeamTrait
+{
+    public function fetchTeam()
+    {
+        return $this->fetch(\App\Models\Team::class);
+    }
+    
+}
